Hand the DAC back on teardown, and fix bypass position and ring wrap
Four defects in the exclusive USB path, one of them visible in a field log.

Disengage bypass when the driver releases the DAC. An unplug or an
exclusive-mode toggle tears the stream down on UsbExclusiveController's
IO thread, but handleBuffer only ever checked bypassActive, so the sink
kept writing to a closed driver until the 400 ms wedge watchdog noticed.
A field log shows the cost: the device closed at 09:44:51.365 and the
watchdog did not trip until 51.759 — 394 ms of silence and 80 "no active
stream format" JNI error lines, one per renderer tick. Both driver flags
are cleared only by an explicit close()/stop(), so checking them hands
the stream to the delegate on the very next buffer instead.

Report CURRENT_POSITION_NOT_SET, not C.TIME_UNSET. They differ by one
(Long.MIN_VALUE vs Long.MIN_VALUE + 1) and MediaCodecAudioRenderer only
filters the former; the value that slips past is then assigned straight
through while allowPositionDiscontinuity is set, which onPositionReset()
does on every seek — exactly when startTimeUs is back to unset. The
position jumped to Long.MIN_VALUE + 1 after every seek in bypass.

Track pause state unconditionally. `paused` mirrors the transport, not
bypass, but both setters sat behind `if (bypassActive)`. If bypass
dropped while paused — DAC unplugged, exclusive mode off, watchdog trip
— the matching play() no-opped too and the gate stayed up, so the next
engage refused every buffer forever with no log and no way back short
of reset().

Widen the ring cursors to uint64_t. Sizing the ring to ~250 ms dropped
the power-of-two invariant in favour of modulo indexing, but size_t is
32 bits on armeabi-v7a, which we ship, and no size start() picks is a
power of two: 2^32 % ringBytes_ is non-zero at every supported format
and not even frame-aligned at 96k/24-bit. Producer and consumer then
address different bytes until both have wrapped — one ring depth of
scrambled PCM, after ~6.8 h at 44.1k and ~23 min at 384k/32-bit/2ch.
The old mask wrapped cleanly; 64-bit counters restore that and put the
wrap ~190,000 years out. Also drops a dead ringSize() helper whose
comment claimed the wrap was already handled.

+++ usb/libusb_uac_driver.h
@@ -237,8 +237,17 @@ private:
     // consumer is the event thread via onIso → drainRing.
     std::vector<uint8_t> ring_;
     size_t ringBytes_ = 1u << 20;  // resized at start() to ~250 ms of audio
-    std::atomic<size_t> ringHead_{0};  // producer cursor (writePcm)
-    std::atomic<size_t> ringTail_{0};  // consumer cursor (onIso)
+    // uint64_t, not size_t: with modulo indexing the cursors must not wrap at
+    // a boundary that isn't a multiple of ringBytes_, or the producer and the
+    // consumer map the same logical byte to different offsets until both have
+    // wrapped — one ring depth of scrambled PCM. size_t is 32 bits on
+    // armeabi-v7a, which we ship, and no size start() picks is a power of two,
+    // so 2^32 % ringBytes_ != 0 at every supported format (and isn't even
+    // frame-aligned at 96k/24-bit). That wrap lands after ~6.8 h at 44.1k and
+    // ~23 min at 384k/32-bit/2ch. The old power-of-two mask wrapped cleanly;
+    // 64-bit counters restore that for free (~190,000 years at 384k).
+    std::atomic<uint64_t> ringHead_{0};  // producer cursor (writePcm)
+    std::atomic<uint64_t> ringTail_{0};  // consumer cursor (onIso)
 
     mutable std::mutex mutex_;          // guards open/start/stop only
     libusb_context* ctx_ = nullptr;
